[fp8] refactor hook
