+++ colossalai/quantization/fp8_hook.py
